+++ starters/eslintrc.js
@@ -1,4 +1,4 @@
-module.exports = require('hxmstyle')({
+module.exports = require('@hugsmidjan/hxmstyle')({
   // Place your project-specific additions or overrides here
   // using standard ESLint config syntax...