+++ starters/prettierrc.js
@@ -1,4 +1,4 @@
-module.exports = require('hxmstyle/prettier')({
+module.exports = require('@hugsmidjan/hxmstyle/prettier')({
   // Place your project-specific additions or overrides here
   // using standard Prettierrc syntax...
 });